Be more persistent about finding a home trash dir
Instead of giving up creating a home trash dir in the Trash constructor, try again every time a home trash dir is needed.  The map doesn't get a nullptr for the home trash device.  Instead of storing a pointer to the home trash dir, which may or may not exist, store the path where it should be located, and the device of that filesystem.  These are used every time a home trash dir is needed to check in the map, and then to try and create it again.  This obviously shouldn't happen, but if it does it gives the user a chance to fix things without shutting down the whole application.

Expand the definition of a "trash dir", a directory which cannot be trashed, to include not just all the descendants of any trash root directory, but also any ancestor.  Usually, the only ancestor of a trash directory will be the root for that filesystem so it won't often be getting trashed, but this also identifies .Trash as well as the actual .Trash/$UID and .Trash-$UID.  Directories are also identified if they do not contain the "files" and "info" directories, are not "accessible".  However, invalid $TOPDIR/.Trash (not a directory, or no sticky bit) will not be considered a trash dir.

Also slightly change the definition of a valid main trash directory to not require S_IXOTH to be set, although any trash operations are likely to fail with permission denied.