Fix crash from refreshing files in pseudo-dirs
The Refresher can call DirTree:refresh() with a file in a pseudo-dir.  This causes an attempt to unattic the parent of the file, which is an Attic.  Chase further up the tree until an actual directory is found.
Move prepareForRefresh from SelectionModel to CleanupCollection (unnamed namespace), the only caller.
Provide a SysUtil function fo getting the parent directory of a path.