Trash copy-and-delete support
Respect the option to only use the home trash directory.  This may be useful to prevent the creation of .Trash-1000 directories at the roots of devices where people don't want them, although it may cause more cross-device moves.

Implement copy-and-delete, so that trash moves where rename() fails with EXDEV (invalid cross-device link) can be performed by copying the item and then deleting it.  QFile::rename does this automatically for files but not for directories, C does not offer a function at all, and C++ not until C++17.  So use external processes to copy and delete.  Copying directories in particular may fail for reasons that are hard to identify in advance, so revert failed copies and even failed deletes.

Because copying between different filesystems (not strictly cross-device because it can happen for bind mounts on the same device) can be slow, and even has a small risk of leaving the Trash in an inconsistent state if it is interrupted, check the default-false option before going ahead.  Restores will always attempt copy-and-delete if necessary because presumably that's how they go to be on different filesystems.