Hopefully the last tooltip cleanup
Fix UnreadableDirsWindow to not display paths as rich text.

Change FormatUtil::pathTooltip to not modify its input.  This is too error-prone, unexpectedly changing pathnames.  Most callers have to make their own copy anyway; the one that doesn't is not greatly impacted by pathTooltip() taking and returning a copy of the original path.