Avoid repetitive failures
For each moveToTrash operation, just try (and possibly fail) to locate/create a trash directory once for each device number.  To do this, store a null pointer in the map if there is a complete failure including the home trash directory.  When falling back to the home directory, put an entry in the map for that device and the home TrashDir.  Then prune out the null/temporary entries from the map after the entire operation is complete.  This avoids writing hundreds or thousands of error messages to the log (and inevitably making things slower) when there is almost no hope of success after the first one in a single batch, but still allows for the application to recover at the next attempt if a missing trash directory is created or permissions are changed.  The application will still fail permanently (and spam the log about it) if a known TrashDir directory is removed; this can be solved if it seems worthwhile.

Improve some of the other Trash error handling and messaging.